scripts/kafka_scripts/requirements.txt
confluent-kafka==2.3.0
snowflake-connector-python
python-dotenv